pyproject.toml
[project]
name = "zirk-rh-financeiro"
version = "0.1.0"
description = "Plataforma de RH e financeiro desenvolvida em Django para a Zirk Marcenaria."
authors = [
    {name = "BernardoVieiraDev",email = "bernardovdev@gmail.com"}
]
license = {text = "Proprietary"}
requires-python = ">=3.12"
dependencies = [
    "django",
    "xlsxwriter",
    "requests",
    "cryptography (>=46.0.3,<47.0.0)",
    "djfernet (>=0.8.1,<1.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"